# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /operator/vaults:
    get:
      operationId: listOperatorVaults
      tags: [operator]
      summary: List the signed-in host's own vaults for the dashboard; never prediction content
      responses:
        "200":
          description: Vault cards only, deleted vaults excluded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperatorVaultList" }
        "401": { description: Authentication required }
    post:
      operationId: createOperatorVault
      tags: [operator]
      summary: Spend an unspent entitlement to create a draft vault
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateDraftVaultInput" } } }
      responses:
        "201": { description: Draft vault created, content: { application/json: { schema: { $ref: "#/components/schemas/CreateDraftVaultResult" } } } }
        "400": { description: Invalid vault type or subject values }
        "401": { description: Authentication required }
        "409": { description: "The entitlement is missing, already spent, or not paid" }
  /operator/vault-types:
    get:
      operationId: listOperatorVaultTypes
      tags: [operator]
      summary: List the active vault types a host can choose from when creating a vault
      responses:
        "200": { description: Active vault types, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/OperatorVaultType" } } } } }
        "401": { description: Authentication required }
  /operator/vaults/{vaultId}/delete:
    post:
      operationId: deleteOperatorVault
      tags: [operator]
      summary: Host-initiated soft delete; the vault and its data are retained but hidden
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Vault marked deleted, content: { application/json: { schema: { $ref: "#/components/schemas/OperatorVaultDeleteResult" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found, not owned by this host, or already deleted" }
  /operator/vaults/{vaultId}/setup:
    get:
      operationId: getVaultSetupDetail
      tags: [operator]
      summary: Get a draft vault's current setup state
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Setup detail, content: { application/json: { schema: { $ref: "#/components/schemas/VaultSetupDetail" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
    patch:
      operationId: updateVaultSetup
      tags: [operator]
      summary: Autosave a draft vault's setup fields (event date, schedule, milestone)
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateVaultSetupInput" } } }
      responses:
        "200": { description: Setup saved, content: { application/json: { schema: { $ref: "#/components/schemas/VaultSetupDetail" } } } }
        "400": { description: "Invalid schedule for this plan, or vault is not a draft" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/cover:
    post:
      operationId: uploadVaultCover
      tags: [operator]
      summary: Upload a vault's cover photo (paid tiers only); replaces any existing photo
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Cover uploaded, content: { application/json: { schema: { $ref: "#/components/schemas/VaultCoverInfo" } } } }
        "400": { description: "File is not a valid JPEG/PNG, or exceeds the size limit" }
        "401": { description: Authentication required }
        "403": { description: Lockbox vaults cannot upload a cover photo }
        "404": { description: "Vault not found or not owned by this host" }
    delete:
      operationId: removeVaultCover
      tags: [operator]
      summary: Remove a vault's uploaded cover photo, reverting to the vault type's silhouette
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Cover removed, content: { application/json: { schema: { $ref: "#/components/schemas/VaultCoverInfo" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/guest-layout:
    patch:
      operationId: updateVaultGuestLayout
      tags: [operator]
      summary: Change how guests see the prompt list; allowed at any vault status
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/GuestLayoutInput" } } }
      responses:
        "200": { description: Guest layout saved, content: { application/json: { schema: { $ref: "#/components/schemas/GuestLayoutInfo" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/seal-readiness:
    get:
      operationId: getSealReadiness
      tags: [operator]
      summary: Check whether a draft vault meets every condition to be sealed
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Seal readiness, content: { application/json: { schema: { $ref: "#/components/schemas/SealReadiness" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/seal:
    post:
      operationId: sealVaultAction
      tags: [operator]
      summary: Seal a draft vault, locking its prompts and computing its reveal schedule
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Vault sealed, content: { application/json: { schema: { $ref: "#/components/schemas/SealVaultResult" } } } }
        "400": { description: "Vault does not yet meet every condition to be sealed" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found, not owned by this host, or already sealed" }
  /operator/vaults/{vaultId}/setup/schedule-preview:
    post:
      operationId: previewVaultSchedule
      tags: [operator]
      summary: Preview reveal dates for a chosen schedule before sealing
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SchedulePreviewInput" } } }
      responses:
        "200": { description: Reveal date preview, content: { application/json: { schema: { $ref: "#/components/schemas/SchedulePreviewResult" } } } }
        "400": { description: "Invalid schedule, plan, or milestone date" }
        "401": { description: Authentication required }
  /operator/vaults/{vaultId}/prompts:
    get:
      operationId: listVaultPrompts
      tags: [operator]
      summary: List a vault's prompts, custom prompts first, in display order
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Prompt list, content: { application/json: { schema: { $ref: "#/components/schemas/VaultPromptList" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
    post:
      operationId: addCustomPrompt
      tags: [operator]
      summary: Add a custom free-text prompt to a draft vault
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AddCustomPromptInput" } } }
      responses:
        "201": { description: Prompt added, content: { application/json: { schema: { $ref: "#/components/schemas/VaultPrompt" } } } }
        "400": { description: "Prompt is empty, too long, or vault is not a draft" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/prompts/order:
    put:
      operationId: reorderVaultPrompts
      tags: [operator]
      summary: Reorder a draft vault's prompts by drag-to-reorder
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ReorderPromptsInput" } } }
      responses:
        "200": { description: New order saved, content: { application/json: { schema: { $ref: "#/components/schemas/VaultPromptList" } } } }
        "400": { description: "List does not match this vault's prompts, or vault is not a draft" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found or not owned by this host" }
  /operator/vaults/{vaultId}/prompts/{vaultQuestionId}:
    patch:
      operationId: toggleVaultPrompt
      tags: [operator]
      summary: Turn a draft vault's prompt on or off
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - $ref: "#/components/parameters/VaultQuestionId"
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/TogglePromptInput" } } }
      responses:
        "200": { description: Prompt updated, content: { application/json: { schema: { $ref: "#/components/schemas/VaultPrompt" } } } }
        "400": { description: Vault is not a draft }
        "401": { description: Authentication required }
        "404": { description: "Vault or prompt not found" }
  /operator/vaults/{vaultId}/date-change:
    get:
      operationId: getSealedVaultDateInfo
      tags: [operator]
      summary: Get a sealed vault's current event date, schedule, and lock state
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      responses:
        "200": { description: Date info, content: { application/json: { schema: { $ref: "#/components/schemas/SealedVaultDateInfo" } } } }
        "401": { description: Authentication required }
        "404": { description: "Vault not found, not owned by this host, or not sealed" }
    post:
      operationId: changeSealedVaultEventDate
      tags: [operator]
      summary: Move a sealed vault's event date and recalculate its reveal schedule
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SealedVaultDateChangeInput" } } }
      responses:
        "200": { description: Date changed and reveals recalculated, content: { application/json: { schema: { $ref: "#/components/schemas/SealedVaultDateChangeResult" } } } }
        "400": { description: "Date is locked, or would push a reveal into the past or today" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found, not owned by this host, or not sealed" }
  /operator/vaults/{vaultId}/date-change/preview:
    post:
      operationId: previewSealedVaultDateChange
      tags: [operator]
      summary: Preview a sealed vault's recalculated reveal dates before confirming
      parameters: [{ $ref: "#/components/parameters/VaultId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SealedVaultDateChangeInput" } } }
      responses:
        "200": { description: Recalculated preview, content: { application/json: { schema: { $ref: "#/components/schemas/SchedulePreviewResult" } } } }
        "400": { description: "Date is locked, or would push a reveal into the past or today" }
        "401": { description: Authentication required }
        "404": { description: "Vault not found, not owned by this host, or not sealed" }
...
components:
  parameters:
    VaultId:
      name: vaultId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    CreateDraftVaultInput:
      type: object
      required: [billingRecordId, vaultTypeId, name]
      properties:
        billingRecordId: { type: string, format: uuid }
        vaultTypeId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        subjectValues: { type: object, additionalProperties: { type: string } }
    CreateDraftVaultResult:
      type: object
      required: [vaultId]
      properties:
        vaultId: { type: string, format: uuid }
    OperatorVaultType:
      type: object
      required: [id, slug, name, requiredSubjectTokens]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        requiredSubjectTokens: { type: array, items: { type: string } }
    VaultSetupDetail:
      type: object
      required: [id, name, status, planTier, entitledPlanTier, vaultTypeId, vaultTypeName, vaultTypeSlug, revealSchedule, anchorDate, milestoneDate, milestoneLabel, timeZone, coverObjectKey, guestLayout, guestToken, subjectValues]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [draft, sealed, active, completed] }
        planTier: { type: string, enum: [lockbox, safe, vault, deep_vault] }
        entitledPlanTier: { type: string, enum: [lockbox, safe, vault, deep_vault] }
        vaultTypeId: { type: string, format: uuid }
        vaultTypeName: { type: string }
        vaultTypeSlug: { type: string }
        revealSchedule: { type: ["string", "null"], enum: [weekly_sprint, monthly_x3, monthly_year, half_then_annual, annual_keepsake, null] }
        anchorDate: { type: ["string", "null"], format: date }
        milestoneDate: { type: ["string", "null"], format: date }
        milestoneLabel: { type: ["string", "null"], maxLength: 60 }
        timeZone: { type: ["string", "null"] }
        coverObjectKey: { type: ["string", "null"] }
        guestLayout: { type: string, enum: [one_at_a_time, all_prompts] }
        subjectValues:
          type: object
          additionalProperties: { type: string }
          description: The vault's subject-name token values, keyed by the literal bracketed token (e.g. "[Baby]"), as the host entered them.
        guestToken:
          type: string
          description: >-
            The raw guest link token. Only ever returned to this vault's own
            authenticated host on this read; never exposed to a guest or an
            unauthenticated request.
    SealReadiness:
      type: object
      required: [ready, reasons, promptCount, scheduleName, tierName, timeZoneLabel, firstRevealDate, lastRevealDate]
      properties:
        ready: { type: boolean }
        reasons: { type: array, items: { type: string } }
        promptCount: { type: integer }
        timeZoneLabel: { type: ["string", "null"], description: "The chosen time zone's exact dropdown label, or null if none is chosen yet." }
        scheduleName: { type: ["string", "null"], enum: [weekly_sprint, monthly_x3, monthly_year, half_then_annual, annual_keepsake, null] }
        tierName: { type: string, enum: [lockbox, safe, vault, deep_vault] }
        firstRevealDate: { type: ["string", "null"], format: date }
        lastRevealDate: { type: ["string", "null"], format: date }
    SealVaultResult:
      type: object
      required: [vaultId, guestToken, sealedAt]
      properties:
        vaultId: { type: string, format: uuid }
        guestToken: { type: string }
        sealedAt: { type: string, format: date-time }
    UpdateVaultSetupInput:
      type: object
      required: [planTier]
      properties:
        planTier: { type: string, enum: [lockbox, safe, vault, deep_vault] }
        revealSchedule: { type: ["string", "null"], enum: [weekly_sprint, monthly_x3, monthly_year, half_then_annual, annual_keepsake, null] }
        anchorDate: { type: ["string", "null"], format: date }
        milestoneDate: { type: ["string", "null"], format: date }
        milestoneLabel: { type: ["string", "null"], maxLength: 60 }
        timeZone: { type: ["string", "null"] }
...
    SchedulePreviewResult:
      type: object
      required: [revealSlots]
      properties:
        revealSlots: { type: array, items: { $ref: "#/components/schemas/RevealSlotPreview" } }
...
    VaultPromptList:
      type: object
      required: [prompts]
      properties:
        prompts: { type: array, items: { $ref: "#/components/schemas/VaultPrompt" } }
...
    VaultCoverInfo:
      type: object
      required: [coverObjectKey]
      properties:
        coverObjectKey: { type: ["string", "null"], description: "Object storage path for the uploaded cover photo, or null to use the vault type's silhouette." }
    GuestLayoutInput:
      type: object
      required: [guestLayout]
      properties:
        guestLayout: { type: string, enum: [one_at_a_time, all_prompts] }
    GuestLayoutInfo:
      type: object
      required: [guestLayout]
      properties:
        guestLayout: { type: string, enum: [one_at_a_time, all_prompts] }
...
    OperatorVaultList:
      type: object
      required: [vaults]
      properties:
        vaults:
          type: array
          items:
            type: object
            required: [id, name, status, cardStatus, createdAt, vaultTypeName, vaultTypeSlug]
            properties:
              id: { type: string, format: uuid }
              name: { type: string }
              status: { type: string, enum: [draft, sealed, active, completed] }
              cardStatus: { type: string, enum: [draft, sealed, partially_unlocked, fully_unlocked, completed], description: "Display status; active splits into partially_unlocked or fully_unlocked." }
              createdAt: { type: string, format: date-time }
              vaultTypeName: { type: string }
              vaultTypeSlug: { type: string }
    OperatorVaultDeleteResult:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [deleted] }
...